Added incomplete database support for Catalists.
Schema for Catalists: title, list-id, contents--contents is a list of
DynamicEmbeddedDocument instances that will hold user-generated
key-value pairs. Desired behavior is that across a single Catalist, all
elements (instances of DynamicEmbeddedDocument) have the *same schema*.
We want this schema to be dynamic, but shared by all the elements. Not
sure if current implementation does this (in fact, it probably doesn't.)